AP_Scripting: introduce serial device support
Allows a script to be a device from the perspective of any serial
protocol. The script can read and write data and have the protocol
handle it as if exchanged over a physical serial port. The script can
then do protocol translation, data filtering and validation,
hardware-in-the-loop simulation, experimentation, etc., especially in
combination with the scripting protocol which lets the script itself
exchange data over a physical serial port.

This just sets up basic bindings and buffers for now.